project2/requirements.txt
Flask
Flask-SocketIO
werkzeug.utils
